Move arrayToTexture function
Make arrayToTexture accesbile from gammacv module scope, add tests for it.
