+++ lib/ops/conv2d/kernels.js
@@ -1,5 +1,5 @@
 import Tensor from '../../program/tensor';
-import { arrayToTexture } from '../../program/utils';
+import { arrayToTexture } from '../../utils';
 
 /**
  * @param {number} kernelSize
