+++ lib/program/utils.js
@@ -91,17 +91,3 @@ export function parallelReductionGetSteps2d(
 export function clacConvolution(inputLength, kernelSide, stride = 1) {
   return Math.ceil(((inputLength - kernelSide) + 1) / stride);
 }
-
-export function arrayToTexture(arr, alpha) {
-  const res = new Array(arr.length * 4);
-
-  for (let i = 0; i < res.length; i += 1) {
-    if ((i + 1) % 4 === 0 && typeof alpha === 'number') {
-      res[i] = alpha;
-    } else {
-      res[i] = arr[~~(i / 4)];
-    }
-  }
-
-  return res;
-}
