+++ lib/utils.js
@@ -49,3 +49,17 @@ export function deprecationWarning(name, msg) {
 export function deprecationError(name, msg) {
   throw new DeprecationError(`GammaCV Deprecation Error: "${name}" is deprecated${msg ? `, ${msg}` : ''}. "${name}" and was removed.`);
 }
+
+export function arrayToTexture(arr, alpha) {
+  const res = new Array(arr.length * 4);
+
+  for (let i = 0; i < res.length; i += 1) {
+    if ((i + 1) % 4 === 0 && typeof alpha === 'number') {
+      res[i] = alpha;
+    } else {
+      res[i] = arr[~~(i / 4)];
+    }
+  }
+
+  return res;
+}
