+++ test/ops/conv2d.spec.js
@@ -1,8 +1,7 @@
 import testImage from '../assets/lena.png';
 import * as gm from '../../lib';
-import { arrayToTexture } from '../../lib/program/utils';
 
-describe.only('Convolution', () => {
+describe('Convolution', () => {
   let sess;
   const setup = (input, kernel, factor, bias) => {
     const op = gm.conv2d(input, kernel, factor, bias);
@@ -26,7 +25,7 @@ describe.only('Convolution', () => {
 
   it('Identity kernel', async () => {
     const input = await gm.imageTensorFromURL(testImage);
-    const identityKernel = new gm.Tensor('float32', [3, 3, 4], new Float32Array(arrayToTexture([
+    const identityKernel = new gm.Tensor('float32', [3, 3, 4], new Float32Array(gm.arrayToTexture([
       0, 0, 0,
       0, 1, 0,
       0, 0, 0,
@@ -39,17 +38,17 @@ describe.only('Convolution', () => {
   });
 
   it('Identity kernel with factor and bias', async () => {
-    const input = new gm.Tensor('float32', [3, 3, 4], new Float32Array(arrayToTexture([
+    const input = new gm.Tensor('float32', [3, 3, 4], new Float32Array(gm.arrayToTexture([
       1, 2, 3,
       4, 5, 6,
       7, 8, 9,
     ])));
-    const identityKernel = new gm.Tensor('float32', [3, 3, 4], new Float32Array(arrayToTexture([
+    const identityKernel = new gm.Tensor('float32', [3, 3, 4], new Float32Array(gm.arrayToTexture([
       0, 0, 0,
       0, 1, 0,
       0, 0, 0,
     ])));
-    const predictedOutput = new gm.Tensor('float32', [3, 3, 4], new Float32Array(arrayToTexture([
+    const predictedOutput = new gm.Tensor('float32', [3, 3, 4], new Float32Array(gm.arrayToTexture([
       0.8, 1.3, 1.8,
       2.3, 2.8, 3.3,
       3.8, 4.3, 4.8,
@@ -63,7 +62,7 @@ describe.only('Convolution', () => {
 
   describe('buitin kernels', () => {
     it('gaussianBlur', () => {
-      const input = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(arrayToTexture([
+      const input = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(gm.arrayToTexture([
         10, 20, 30,
         40, 50, 60,
         70, 80, 90,
@@ -73,7 +72,7 @@ describe.only('Convolution', () => {
 
       sess.runOp(test.op, 0, test.output);
 
-      const predictedOutput = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(arrayToTexture([
+      const predictedOutput = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(gm.arrayToTexture([
         21, 28, 35,
         43, 50, 57,
         65, 72, 79,
@@ -82,7 +81,7 @@ describe.only('Convolution', () => {
       gm.assert(gm.tensorAssertCloseEqual(test.output, predictedOutput, 0.003), 'Input and output should be close equal');
     });
     it('Box Blur', () => {
-      const input = new gm.Tensor('float32', [3, 3, 4], new Float32Array(arrayToTexture([
+      const input = new gm.Tensor('float32', [3, 3, 4], new Float32Array(gm.arrayToTexture([
         1, 2, 3,
         4, 5, 6,
         7, 8, 9,
@@ -92,7 +91,7 @@ describe.only('Convolution', () => {
 
       sess.runOp(test.op, 0, test.output);
 
-      const predictedOutput = new gm.Tensor('float32', [3, 3, 4], new Float32Array(arrayToTexture([
+      const predictedOutput = new gm.Tensor('float32', [3, 3, 4], new Float32Array(gm.arrayToTexture([
         2.333, 3, 3.666,
         4.333, 5, 5.666,
         6.333, 7, 7.666,
@@ -102,7 +101,7 @@ describe.only('Convolution', () => {
     });
 
     it('Sharpen', () => {
-      const input = new gm.Tensor('float32', [3, 3, 4], new Float32Array(arrayToTexture([
+      const input = new gm.Tensor('float32', [3, 3, 4], new Float32Array(gm.arrayToTexture([
         1, 2, 3,
         4, 5, 6,
         7, 8, 9,
@@ -112,7 +111,7 @@ describe.only('Convolution', () => {
 
       sess.runOp(test.op, 0, test.output);
 
-      const predictedOutput = new gm.Tensor('float32', [3, 3, 4], new Float32Array(arrayToTexture([
+      const predictedOutput = new gm.Tensor('float32', [3, 3, 4], new Float32Array(gm.arrayToTexture([
         -3, -1, 1,
         3, 5, 7,
         9, 11, 13,
@@ -122,7 +121,7 @@ describe.only('Convolution', () => {
     });
 
     it('Invert', () => {
-      const input = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(arrayToTexture([
+      const input = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(gm.arrayToTexture([
         1, 2, 3,
         4, 5, 6,
         7, 8, 9,
@@ -132,7 +131,7 @@ describe.only('Convolution', () => {
 
       sess.runOp(test.op, 0, test.output);
 
-      const predictedOutput = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(arrayToTexture([
+      const predictedOutput = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(gm.arrayToTexture([
         254, 253, 252,
         251, 250, 249,
         248, 247, 246,
@@ -142,7 +141,7 @@ describe.only('Convolution', () => {
     });
 
     it('Edge detection', () => {
-      const input = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(arrayToTexture([
+      const input = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(gm.arrayToTexture([
         10, 0, 30,
         15, 0, 30,
         17, 9, 249,
@@ -152,7 +151,7 @@ describe.only('Convolution', () => {
 
       sess.runOp(test.op, 0, test.output);
 
-      const predictedOutput = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(arrayToTexture([
+      const predictedOutput = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(gm.arrayToTexture([
         0, 0, 0,
         2, 212, 210,
         7, 217, 210,
@@ -162,7 +161,7 @@ describe.only('Convolution', () => {
     });
 
     it('Unsharp Masking', () => {
-      const input = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(arrayToTexture([
+      const input = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(gm.arrayToTexture([
         1, 2, 3,
         4, 5, 6,
         7, 8, 9,
@@ -172,7 +171,7 @@ describe.only('Convolution', () => {
 
       sess.runOp(test.op, 0, test.output);
 
-      const predictedOutput = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(arrayToTexture([
+      const predictedOutput = new gm.Tensor('uint8', [3, 3, 4], new Uint8Array(gm.arrayToTexture([
         0, 1, 2,
         4, 5, 6,
         8, 9, 10,
